Add `importCommonJS` option to `babel-plugin-transform-es2015-modules-commonjs`.
Normally, Babel assumes that ES module code will be importing ES module code
transpiled to CommonJS when the importing code is transpiled to CommonJS.  This
assumption is not necessarily true.  For instance, the imported code could have
been originally written in CommonJS, in which case the code's `default` export
would be the `exports` object (in Node.js), rather than `exports.default` in the
conventional case for transpiled code.  Hence, transforming reference
identifiers of the imported default binding to member expressions accessing
`exports.default` is sometimes incorrect.

Normally, `interopRequireDefault` works around this issue by setting
non-`__esModule`-tagged `exports` to a `default` key of an object.  But
disabling `interopRequireDefault` with the `noInterop` option eliminates the
workaround.

Instead of assuming that imported code is an ES module transpiled to CommonJS,
set the `importCommonJS` option to a truthy value to assume that imported code
is "just CommonJS."